+++ src/tokens.pas
@@ -29,8 +29,8 @@ type
     symbol =
         (nul, ident, number, plus, minus, times, slash, oddsym,
         eql, neq, lss, leq, gtr, geq, lparen, rparen, comma, semicolon,
-        period, becomes, beginsym, endsym, ifsym, thensym,
-        whilesym, dosym, callsym, constsym, varsym, procsym);
+        period, colon, becomes, beginsym, endsym, ifsym, thensym,
+        whilesym, dosym, callsym, constsym, varsym, procsym, integersym);
     symbolSet = set of symbol;
 
     identifier = packed array [1 .. maxIdentLength] of char;