+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project is expected to follow [Semantic Versioning](https://semver.org/).
 
+## [Unreleased]
+
+- Added support for type declaration in `var` and `const`, mandatory and only `integer` datatype for now.
+
 ## [v0.0.3] - 2026-05-22
 
 ### Added